[Refactoring ⚙️] 멤버 정보 전역의 Avatar 카드로 가져옴
